MS-Access import

+++ fnpdbc-dbconvert/src/main/java/dbconvert/software/XConverter.java
@@ -413,7 +413,9 @@ public class XConverter extends BasicSoft implements IDatabaseFactory {
 
 		if (isOutputFileOpen) {
 			dbOut.closeFile();
-			dbOut.deleteFile();
+			if (!myExportFile.isAppend() || dbOut.hasBackup()) {
+				dbOut.deleteFile();
+			}
 			isOutputFileOpen = false;
 		}
 	}
